+++ src/GauntletLayer/GauntletEdits/05_lava.cpp
@@ -21,9 +21,17 @@ void RedesignedGauntletLayer::editLavaGauntlet() {
             bgParticles->setZOrder(-98);
             this->addChild(bgParticles);
         }
-        // ash
+        // glow
         auto bgParticleNode2 = CCParticleSystemQuad::create();
         if (bgParticleNode2) {
+            CCParticleSystemQuad* bgParticles = GameToolbox::particleFromString("60a-1a2a0.45a5a-90a0a107a0a350a0a0a500a0a0a0a0a750a5a90a47a0.811765a0a0.490196a0a0.211765a0a0.5a0a750a0a90a47a0.337255a0a0a0a0a0a0.4a0a0.25a0a0.4a0a0a0a0a0a0a0a0a2a1a0a0a0a26a0a4.5a0a0a0a0a0a0a0a0a0a0a0a0;", NULL, false);
+            bgParticles->setPosition(ccp(winSize.width / 2, director->getScreenBottom() - 225));
+            bgParticles->setZOrder(-98);
+            this->addChild(bgParticles);
+        }
+        // ash
+        auto bgParticleNode3 = CCParticleSystemQuad::create();
+        if (bgParticleNode3) {
             CCParticleSystemQuad* bgParticles = GameToolbox::particleFromString("25a-1a2a0.45a6a90a20a250a10a300a0a0a-225a0a0a0a0a25a10a0a43a1a0a0.792157a0a0a0a1a0.05a0a0a0a38a1a0a0.152941a0a0a0a0.5a0.15a0a0a0.45a0a0a0a0a0a0a0a0a2a1a0a0a0a28a0a0a0a0a0a0a0a0a0a0a0a0a0a0", NULL, false);
             bgParticles->setPosition(ccp(winSize.width / 2, director->getScreenBottom() - 10));
             bgParticles->setZOrder(-98);